Create Repository DTO and update list endpoint
